refactor: extract roster to unit mapper
