+++ hooks/useAllUnits.ts
@@ -1,35 +1,8 @@
-import {
-  type Army,
-  type GameArmy,
-  type GameUnit,
-  type Unit
-} from 'appdeptus/models'
+import { type Army, type GameArmy } from 'appdeptus/models'
+import { mapRosterToUnits } from 'appdeptus/utils'
 import { useMemo } from 'react'
 
 const useAllUnits = (units: Army['roster'] | GameArmy['roster']) =>
-  useMemo(
-    () =>
-      units.reduce<(Unit | GameUnit)[]>((acc, curr) => {
-        switch (curr.type) {
-          case 'embarked':
-            return [
-              ...acc,
-              curr.transport,
-              ...curr.crew.flatMap((embarked) =>
-                embarked.type === 'team'
-                  ? [embarked.leader, embarked.bodyguard]
-                  : embarked
-              )
-            ]
-
-          case 'team':
-            return [...acc, curr.leader, curr.bodyguard]
-
-          default:
-            return [...acc, curr]
-        }
-      }, []),
-    [units]
-  )
+  useMemo(() => mapRosterToUnits(units), [units])
 
 export default useAllUnits
