+++ utils/index.ts
@@ -1,2 +1,3 @@
 export { default as mapNullToUndefined } from './mapNullToUndefined'
+export { default as mapRosterToUnits } from './mapRosterToUnits'
 export { default as supabase } from './supabase'
